+++ src/layers/DenseLayer.h
@@ -5,21 +5,21 @@
 
 class DenseLayer : public Layer {
    public:
-    DenseLayer(int input_size, int output_size, double (*activation)(double));
+    DenseLayer(int input_size, int output_size, float (*activation)(float));
 
-    std::vector<std::vector<double>> momentum;
-    std::vector<std::vector<double>> variance;
-    double beta1, beta2, eta, epsilon;
+    std::vector<std::vector<float>> momentum;
+    std::vector<std::vector<float>> variance;
+    float beta1, beta2, eta, epsilon;
 
-    std::vector<std::vector<double>> weight_delta;
+    std::vector<std::vector<float>> weight_delta;
 
-    std::vector<double> predict(std::vector<double> input) override;
-    std::vector<double> forwardpropagate(std::vector<double> input) override {
+    std::vector<float> predict(std::vector<float> input) override;
+    std::vector<float> forwardpropagate(std::vector<float> input) override {
         return this->predict(input);
     };
-    void out_errors(std::vector<double> target_vector) override;
-    void backpropagate(Layer* connected_layer, std::vector<double> target_vector) override;
-    void update_weights(std::vector<double> input_data, double learning_rate, int t) override;
+    void out_errors(std::vector<float> target_vector) override;
+    void backpropagate(Layer* connected_layer, std::vector<float> target_vector) override;
+    void update_weights(std::vector<float> input_data, float learning_rate, int t) override;
 };
 
 #endif
