+++ src/layers/NoiseLayer.h
@@ -7,17 +7,17 @@
 
 class NoiseLayer : public Layer {
    public:
-    NoiseLayer(int output_size, double noise_chance, double noise_scale = 0.1);
+    NoiseLayer(int output_size, float noise_chance, float noise_scale = 0.1);
 
-    std::vector<double> forwardpropagate(std::vector<double> input) override;
-    void out_errors(std::vector<double> target_vector) override {
+    std::vector<float> forwardpropagate(std::vector<float> input) override;
+    void out_errors(std::vector<float> target_vector) override {
         throw std::runtime_error("NoiseLayer::out_errors() is not valid");
     };
-    void update_weights(std::vector<double> input_data, double learning_rate) override { return; };
+    void update_weights(std::vector<float> input_data, float learning_rate) override { return; };
 
    private:
-    double noise_chance;
-    double noise_scale;
+    float noise_chance;
+    float noise_scale;
 };
 
 #endif
